requirements.txt
# uncomment next line for installation on a100-1
# conda install pytorch torchvision torchaudio cudatoolkit=11.6 -c pytorch -c conda-forge
tensorboardX==1.8
pytorch-lightning
ninja
hydra-core

g2p-en==2.1.0
inflect==4.1.0
librosa==0.7.2
matplotlib==3.2.2
numba==0.48
numpy==1.23.0
pypinyin==0.39.0
pyworld==0.2.10
PyYAML==5.4.1
scikit-learn==1.1.1
scipy==1.8.1
soundfile==0.10.3.post1
tgt==1.4.4
unidecode==1.1.1